Move embeds out of commands to be reusable with NeatQueue
