+++ src/commands/stats/stats.mts
@@ -17,27 +17,27 @@ import type { ApplicationCommandData, ExecuteResponse } from "../base/base.mjs";
 import { BaseCommand } from "../base/base.mjs";
 import { Preconditions } from "../../base/preconditions.mjs";
 import type { QueueData } from "../../services/discord/discord.mjs";
-import type { BaseMatchEmbed } from "./embeds/base-match-embed.mjs";
-import { AttritionMatchEmbed } from "./embeds/attrition-match-embed.mjs";
-import { CtfMatchEmbed } from "./embeds/ctf-match-embed.mjs";
-import { EliminationMatchEmbed } from "./embeds/elimination-match-embed.mjs";
-import { EscalationMatchEmbed } from "./embeds/escalation-match-embed.mjs";
-import { ExtractionMatchEmbed } from "./embeds/extraction-match-embed.mjs";
-import { FiestaMatchEmbed } from "./embeds/fiesta-match-embed.mjs";
-import { FirefightMatchEmbed } from "./embeds/firefight-match-embed.mjs";
-import { GrifballMatchEmbed } from "./embeds/grifball-match-embed.mjs";
-import { InfectionMatchEmbed } from "./embeds/infection-match-embed.mjs";
-import { KOTHMatchEmbed } from "./embeds/koth-match-embed.mjs";
-import { LandGrabMatchEmbed } from "./embeds/land-grab-match-embed.mjs";
-import { MinigameMatchEmbed } from "./embeds/minigame-match-embed.mjs";
-import { OddballMatchEmbed } from "./embeds/oddball-match-embed.mjs";
-import { SlayerMatchEmbed } from "./embeds/slayer-match-embed.mjs";
-import { StockpileMatchEmbed } from "./embeds/stockpile-match-embed.mjs";
-import { StrongholdsMatchEmbed } from "./embeds/strongholds-match-embed.mjs";
-import { TotalControlMatchEmbed } from "./embeds/total-control-match-embed.mjs";
-import { UnknownMatchEmbed } from "./embeds/unknown-match-embed.mjs";
-import { VIPMatchEmbed } from "./embeds/vip-match-embed.mjs";
-import { SeriesMatchesEmbed } from "./embeds/series-matches-embed.mjs";
+import type { BaseMatchEmbed } from "../../embeds/base-match-embed.mjs";
+import { AttritionMatchEmbed } from "../../embeds/attrition-match-embed.mjs";
+import { CtfMatchEmbed } from "../../embeds/ctf-match-embed.mjs";
+import { EliminationMatchEmbed } from "../../embeds/elimination-match-embed.mjs";
+import { EscalationMatchEmbed } from "../../embeds/escalation-match-embed.mjs";
+import { ExtractionMatchEmbed } from "../../embeds/extraction-match-embed.mjs";
+import { FiestaMatchEmbed } from "../../embeds/fiesta-match-embed.mjs";
+import { FirefightMatchEmbed } from "../../embeds/firefight-match-embed.mjs";
+import { GrifballMatchEmbed } from "../../embeds/grifball-match-embed.mjs";
+import { InfectionMatchEmbed } from "../../embeds/infection-match-embed.mjs";
+import { KOTHMatchEmbed } from "../../embeds/koth-match-embed.mjs";
+import { LandGrabMatchEmbed } from "../../embeds/land-grab-match-embed.mjs";
+import { MinigameMatchEmbed } from "../../embeds/minigame-match-embed.mjs";
+import { OddballMatchEmbed } from "../../embeds/oddball-match-embed.mjs";
+import { SlayerMatchEmbed } from "../../embeds/slayer-match-embed.mjs";
+import { StockpileMatchEmbed } from "../../embeds/stockpile-match-embed.mjs";
+import { StrongholdsMatchEmbed } from "../../embeds/strongholds-match-embed.mjs";
+import { TotalControlMatchEmbed } from "../../embeds/total-control-match-embed.mjs";
+import { UnknownMatchEmbed } from "../../embeds/unknown-match-embed.mjs";
+import { VIPMatchEmbed } from "../../embeds/vip-match-embed.mjs";
+import { SeriesMatchesEmbed } from "../../embeds/series-matches-embed.mjs";
 
 export class StatsCommand extends BaseCommand {
   data: ApplicationCommandData = {
