+++ src/embeds/attrition-match-embed.mts
