+++ src/embeds/elimination-match-embed.mts
