+++ src/embeds/stockpile-match-embed.mts
