+++ src/embeds/tests/oddball-match-embed.test.mts
@@ -1,11 +1,11 @@
 import { describe, beforeEach, expect, it } from "vitest";
-import { matchStats, playerXuidsToGametags } from "../../../../services/halo/fakes/data.mjs";
+import { matchStats, playerXuidsToGametags } from "../../services/halo/fakes/data.mjs";
 import { OddballMatchEmbed } from "../oddball-match-embed.mjs";
-import type { HaloService } from "../../../../services/halo/halo.mjs";
-import { Preconditions } from "../../../../base/preconditions.mjs";
-import { aFakeHaloServiceWith } from "../../../../services/halo/fakes/halo.fake.mjs";
-import { aFakeDiscordServiceWith } from "../../../../services/discord/fakes/discord.fake.mjs";
-import type { DiscordService } from "../../../../services/discord/discord.mjs";
+import type { HaloService } from "../../services/halo/halo.mjs";
+import { Preconditions } from "../../base/preconditions.mjs";
+import { aFakeHaloServiceWith } from "../../services/halo/fakes/halo.fake.mjs";
+import { aFakeDiscordServiceWith } from "../../services/discord/fakes/discord.fake.mjs";
+import type { DiscordService } from "../../services/discord/discord.mjs";
 
 const match = Preconditions.checkExists(matchStats.get("cf0fb794-2df1-4ba1-9415-00000oddball"));
