+++ src/embeds/tests/series-matches-embed.test.mts
@@ -1,11 +1,11 @@
 import { describe, beforeEach, expect, it } from "vitest";
-import { matchStats, playerXuidsToGametags } from "../../../../services/halo/fakes/data.mjs";
+import { matchStats, playerXuidsToGametags } from "../../services/halo/fakes/data.mjs";
 import { SeriesMatchesEmbed } from "../series-matches-embed.mjs";
-import type { HaloService } from "../../../../services/halo/halo.mjs";
-import { Preconditions } from "../../../../base/preconditions.mjs";
-import { aFakeHaloServiceWith } from "../../../../services/halo/fakes/halo.fake.mjs";
-import { aFakeDiscordServiceWith } from "../../../../services/discord/fakes/discord.fake.mjs";
-import type { DiscordService } from "../../../../services/discord/discord.mjs";
+import type { HaloService } from "../../services/halo/halo.mjs";
+import { Preconditions } from "../../base/preconditions.mjs";
+import { aFakeHaloServiceWith } from "../../services/halo/fakes/halo.fake.mjs";
+import { aFakeDiscordServiceWith } from "../../services/discord/fakes/discord.fake.mjs";
+import type { DiscordService } from "../../services/discord/discord.mjs";
 
 const ctfMatch = Preconditions.checkExists(matchStats.get("d81554d7-ddfe-44da-a6cb-000000000ctf"));
 const kothMatch = Preconditions.checkExists(matchStats.get("e20900f9-4c6c-4003-a175-00000000koth"));
