+++ src/embeds/tests/strongholds-match-embed.test.mts
@@ -1,11 +1,11 @@
 import { describe, beforeEach, expect, it } from "vitest";
-import { matchStats, playerXuidsToGametags } from "../../../../services/halo/fakes/data.mjs";
+import { matchStats, playerXuidsToGametags } from "../../services/halo/fakes/data.mjs";
 import { StrongholdsMatchEmbed } from "../strongholds-match-embed.mjs";
-import type { HaloService } from "../../../../services/halo/halo.mjs";
-import { Preconditions } from "../../../../base/preconditions.mjs";
-import { aFakeHaloServiceWith } from "../../../../services/halo/fakes/halo.fake.mjs";
-import type { DiscordService } from "../../../../services/discord/discord.mjs";
-import { aFakeDiscordServiceWith } from "../../../../services/discord/fakes/discord.fake.mjs";
+import type { HaloService } from "../../services/halo/halo.mjs";
+import { Preconditions } from "../../base/preconditions.mjs";
+import { aFakeHaloServiceWith } from "../../services/halo/fakes/halo.fake.mjs";
+import type { DiscordService } from "../../services/discord/discord.mjs";
+import { aFakeDiscordServiceWith } from "../../services/discord/fakes/discord.fake.mjs";
 
 const match = Preconditions.checkExists(matchStats.get("099deb74-3f60-48cf-8784-0strongholds"));
